requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-ai>=0.0.14
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
aiohttp>=3.9.0
lxml>=4.9.0
websockets>=12.0
reportlab>=4.0.0
Pillow>=10.0.0
